Improve performance
Use recvmmsg with batch size of 10, this moves the QPS needle (with the
includes reflect server, from 20xK qps, to 320+K qps (apple m2 -
performance mode, asahi linux).

This is likely not yet to be cross platform as I'm abusing ipv4 stuff
which seems to work with ipv6 (on linux).

Tested with:

 dnsperf -s 127.0.0.1 -p 8053 -d queries-files -l 5 -c 4 -T 4
~~~
Statistics:

  Queries sent:         1612047
  Queries completed:    1612047 (100.00%)
  Queries lost:         0 (0.00%)

  Response codes:       NOERROR 1612047 (100.00%)
  Average packet size:  request 32, response 92
  Run time (s):         5.000671
  Queries per second:   322366.138464
~~~

Signed-off-by: Miek Gieben <miek@miek.nl>