ServeMux: test and add zone to context
Add a Zone to the context for the Hander, this was the handler knows
what pattern (=zone) was used to get to it. Idea was stolen from CoreDNS
where there is similar code - this is much smaller, and being here more
useful.

The default lookup for the root zone was wrong in case of DS queries,
this is fixed and tested.

Signed-off-by: Miek Gieben <miek@miek.nl>