Documentation update on the behavior of WriteTo
WriteTo implies that the behavior of net.Conn is
only altered when a *net.TCPConn is the underlying
type of w.

However the implementation is actually a specialization
for *net.UDPConn. This difference means in cases where
net.Conn may be neither (i.e. is nil, or a different type)
then the TCP behavior applies and must be accounted for.