atomdns: Use *Node in zones
@cl: Use *dnszone.Node in zones, makes inplace updates possible

This makes signing much faster as the node is update in place (rrsigs
are added), otherwise we needed 2 loops to add all the sigs (and keep
that list in memory).

Fairly trivial, note that dbsqlite, now sees nil, where previously it
was dnsnode.Node{}s, this might still trigger nil-pointer crashes.

Signed-off-by: Miek Gieben <miek@miek.nl>