fix: adjust content-length to match body len
Closes #520

According to the example provided in section 4.2.2 of RFC 8484 (DoH) the
content-length is equal to the byte length of the HTTP body. The body
does not include the 2 bytes encoding the length of the message as it is
usually done with TCP, because the examples all start with two leading
zeroes, and section 4.1 explicitly states that the query ID should be
zero on DoH queries.

https://www.rfc-editor.org/rfc/rfc8484#section-4.2.2
Signed-off-by: Florian Maury <florian.maury@metempsychose.fr>