dns: ZoneParser.IncludeFS public
@cl: ZoneParser: IncludeFS public instead of SetIncludeFS.

Make fsys field public and remove the setter, this is inline with the
IncludeAllowFunc, but not SetDefaultTTL, the latter is needed because
ttlstate is slightly more complex, allthough we could make that a public
field as well.

There is one Setter, which is SetWriteDeadline(), and some in dnsutil,
but most things are just public fields in struct, move zoneparser 1 step
closer to that too.

Signed-off-by: Miek Gieben <miek@miek.nl>